What's new: 1. Added un implemented Adhar and Secondary kyc forms
Fixes:
1. Fixed Login pending screen
2 Fixed Individual Fields mismatch

+++ lib/models/bureau_check/individual.dart
@@ -10,7 +10,7 @@ class Individual {
   String? product;
   String? enquiryPurpose;
   int? internalRefNumber;
-  int? loanAmount;
+  double? loanAmount;
   String? firstName;
   String? middleName;
   String? lastName;
